Support 'Source' invert option.

+++ radio/src/opentx.h
@@ -439,7 +439,7 @@ inline void getMixSrcRange(const int source, int16_t & valMin, int16_t & valMax,
   }
 }
 
-void applyExpos(int16_t * anas, uint8_t mode, uint8_t ovwrIdx=0, int16_t ovwrValue=0);
+void applyExpos(int16_t * anas, uint8_t mode, int16_t ovwrIdx=0, int16_t ovwrValue=0);
 int16_t applyLimits(uint8_t channel, int32_t value);
 
 void evalInputs(uint8_t mode);
